requirements.txt
# CalledIt Backend - Phase 1 Dependencies
# =========================================

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# WebSocket
python-socketio==5.11.0

# Database
motor==3.6.0                      # Async MongoDB driver
pymongo==4.9.0

# Cache
redis[hiredis]==5.2.0             # Async Redis with C parser

# Validation & Settings
pydantic==2.9.0
pydantic-settings==2.5.0

# Auth
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4
twilio==9.3.0                     # OTP via Twilio

# HTTP Client
httpx==0.28.1                     # Async HTTP (CricAPI, Claude, etc.)

# AI
anthropic==0.43.0                 # Claude API SDK

# ML
onnxruntime==1.20.1               # ML inference (Python 3.13 compatible)
xgboost==2.1.0                    # ML training
numpy==2.1.0
pandas==2.2.3
scikit-learn==1.5.2               # ML training (train_test_split)
onnxmltools==1.12.0               # XGBoost → ONNX conversion
skl2onnx==1.17.0                  # sklearn → ONNX (onnxmltools dependency)

# Background Jobs
arq==0.25.0                       # Async Redis task queue

# Image Generation
Pillow==10.4.0                    # Share card generation

# AWS
boto3==1.35.0                     # S3, SES, etc.

# File uploads
python-multipart==0.0.9

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-httpx>=0.35.0

# --- PHASE 2 (not installed yet) ---
# firebase-admin==6.3.0           # Push notifications
# sentry-sdk[fastapi]==1.40.0     # Error tracking
# razorpay==1.4.1                 # Payments
